+++ CHANGELOG.md
@@ -8,6 +8,7 @@ next
   * Note: this is only "partially" supported the `parent` argument of the Proc will NOT contain the correct attribute parent yet. It will contain a fake class, that will loudly complain about any attempt to use any of its methods.
 * Fixed `Model.example` to properly handle the case when no attributes are defined on the class.
 * `Model#dump` now issues a warning if its contents have keys for attributes not present on the class. The unknown contents are not dumped.
+* `Hash.load` now supports loading any value that responds to `to_hash`.
 
 2.4.0
 ------