+++ CHANGELOG.md
@@ -9,6 +9,7 @@ next
 * Fixed `Model.example` to properly handle the case when no attributes are defined on the class.
 * `Model#dump` now issues a warning if its contents have keys for attributes not present on the class. The unknown contents are not dumped.
 * `Hash.load` now supports loading any value that responds to `to_hash`.
+* `Time`, `DateTime`, and `Date` now all return ISO 8601 formatted values from `.dump` (via calling `iso8601` on the value).
 
 2.4.0
 ------